% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clone_ocs_repo.R
\name{clone_ocs_repo}
\alias{clone_ocs_repo}
\title{Internal Function}
\usage{
clone_ocs_repo(casestudy, outpath = NULL, fork_repo = NA)
}
\description{
This is not to be used directly by the users.
}
\keyword{internal}
