% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_long}
\alias{data_long}
\title{Global CO2 and GDP Wrangled Data}
\format{
A data frame with 63,095 rows and 6 variables:
\describe{
\item{Country}{name of the country}
\item{Year}{the year data is from}
\item{Label}{type of data, either CO2 Emissions or GDP Growth}
\item{Indicator}{data indicator, either Emissions or GDP}
\item{Value}{number value of the data, in thousands of tons for CO2 emissions, in GDP for GDP growth}
\item{Region}{indicates if the country is the United States or in the Rest of the World}
...
}
}
\source{
\url{https://github.com/opencasestudies/ocs-bp-co2-emissions/blob/master/data/wrangled_data.csv}
}
\usage{
data_long
}
\description{
A dataset containing the amount of CO2 emitted by each country
as well as GDP growth per year since 1949. CO2 emissions value
reported in thousands of tonnes.
}
\keyword{datasets}
