% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flex_load_data.R
\name{flex_load_data}
\alias{flex_load_data}
\title{Download CO2 Emissions Data}
\usage{
flex_load_data(outpath = NULL)
}
\arguments{
\item{outpath}{character string, path to the directory where the downloaded
data folder should be saved to.}
}
\value{
Nothing useful is returned, a data folder will be downloaded and
appear in your files.
}
\description{
Download the CO2 Emissions case study data into a new folder in your current working directory.
}
\details{
This function downloads the OCS CO2 emissions case study data
from GitHub and saves it in a new '/data' folder in your current working directory. This makes
it so all of the relevant data needed to follow along the lesson is easily available in a local folder.
}
\examples{
flex_load_data('/Users/michael/Desktop/')

}
