% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_extra_data.R
\name{load_extra_data}
\alias{load_extra_data}
\title{Internal Function}
\usage{
load_extra_data(casestudy, outpath = NULL)
}
\description{
This is not to be used directly by the users.
}
\keyword{internal}
