% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_imported_data.R
\name{load_imported_data}
\alias{load_imported_data}
\title{Download Open Case Study Imported Data}
\usage{
load_imported_data(casestudy)
}
\arguments{
\item{casestudy}{character string, name of the case study to pull data from.
The input name should follow the same naming scheme as the repository on GitHub:

ocs-bp-rural-and-urban-obesity

ocs-bp-air-pollution

ocs-bp-vaping-case-study

ocs-bp-opioid-rural-urban

ocs-bp-RTC-wrangling

ocs-bp-RTC-analysis

ocs-bp-youth-disconnection

ocs-bp-youth-mental-health

ocs-bp-school-shootings-dashboard

ocs-bp-co2-emissions

ocs-bp-diet}
}
\value{
Nothing useful is returned, R objects will be loaded into the global environment.
}
\description{
Load the specified case study imported data into your global environment
to use as you follow along the case study.
}
\details{
This function downloads the Open Case Study imported data
from GitHub and loads it into your global environment as variables.
This makes it so all the imported data are immediately available to be wrangled and analzed.
}
\examples{
load_imported_data('ocs-bp-co2-emissions')

}
