% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_repo.R
\name{load_repo}
\alias{load_repo}
\title{Download Case Study Repository Folder}
\usage{
load_repo(casestudy, outpath = NULL)
}
\arguments{
\item{casestudy}{character string, name of the case study to download repository folder from.
The input name should follow the same naming scheme as the repository on GitHub:

ocs-bp-rural-and-urban-obesity

ocs-bp-air-pollution

ocs-bp-vaping-case-study

ocs-bp-opioid-rural-urban

ocs-bp-RTC-wrangling

ocs-bp-RTC-analysis

ocs-bp-youth-disconnection

ocs-bp-youth-mental-health

ocs-bp-school-shootings-dashboard

ocs-bp-co2-emissions

ocs-bp-diet}

\item{outpath}{character string, path to the directory where the downloaded
repository should be saved to. If NULL, the default, the repo will be saved
to your working directory.}
}
\value{
nothing useful is returned, a new folder will be saved to the
directory specified in the parameter called outpath.
}
\description{
Download the specified case study repository and save it as a new folder
and R project in a local directory.
}
\details{
This function downloads the specified OCS case study repository
from GitHub as a zip file. The function unzips the folder and saves it as
a new R project in a local directory. This makes it so the case study
repository, including all of the code, data, and document files, are downloaded
with a single function.
}
\examples{
load_repo('ocs-bp-co2-emissions')

}
