% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wide_US}
\alias{wide_US}
\title{United States Environment and Economy Wrangled Data}
\format{
A data frame with 35 rows and 8 variables:
\describe{
\item{Country}{name of the country}
\item{Year}{the year data is from}
\item{Region}{region the data comes from, always United States in this case}
\item{Emissions}{co2 emitted that year, in thousands of tonnes}
\item{GDP}{GDP growth that year, in GDP}
\item{Energy}{energy use per person, in kg of oil (or equivalent)}
\item{Disasters}{number of natural disasters that year, integer}
\item{Temperature}{average temperature across the contiguous U.S. from Jan-Dec, in degrees Fahrenheit}
...
}
}
\source{
\url{https://github.com/opencasestudies/ocs-bp-co2-emissions/blob/master/data/wrangled_US_data.csv}
}
\usage{
wide_US
}
\description{
A dataset containing data on CO2 emissions, GDP growth, energy useage,
natural disasters and average temperature in the United States since 1980.
}
\keyword{datasets}
