% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_co2_repo.R
\name{load_co2_repo}
\alias{load_co2_repo}
\title{Load CO2 Emissions Repository}
\usage{
load_co2_repo(outpath = NULL)
}
\arguments{
\item{outpath}{character string, path to the directory where the downloaded
repository should be saved to.}
}
\value{
nothing useful is returned, a new R project will be loaded in the
directory specified in outpath.
}
\description{
Download the CO2 Emissions case study repository and save it as a new R project in a local directory.
Need to have a personal access token (PAT) to work.
}
\details{
This function downloads the OCS CO2 emissions case study repository
on GitHub and saves it as a new R project in a local directory. This makes
it so all of the raw and wrangled data is downloaded using one function.
Need to have a personal access token (PAT) to work.
}
\examples{
load_co2_repo('/Users/michael/Desktop/')

}
