% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_simpler_import.R
\name{load_simpler_import}
\alias{load_simpler_import}
\title{Internal Function}
\usage{
load_simpler_import(casestudy, outpath = NULL)
}
\description{
This is not to be used directly by the users.
}
\keyword{internal}
