% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_wrangled_data.R
\name{load_wrangled_data}
\alias{load_wrangled_data}
\title{Download Open Case Study Wrangled Data}
\usage{
load_wrangled_data(casestudy, outpath = NULL)
}
\arguments{
\item{casestudy}{character string, name of the case study to pull data from.
The input name should follow the same naming scheme as the repository on GitHub:

ocs-bp-rural-and-urban-obesity

ocs-bp-air-pollution

ocs-bp-vaping-case-study

ocs-bp-opioid-rural-urban

ocs-bp-RTC-wrangling (Right-to-Carry Part 1)

ocs-bp-RTC-analysis (Right-to-Carry Part 2)

ocs-bp-youth-disconnection

ocs-bp-youth-mental-health

ocs-bp-school-shootings-dashboard

ocs-bp-co2-emissions

ocs-bp-diet}

\item{outpath}{character string, path to the directory where the downloaded
data folder should be saved to.}
}
\value{
Nothing useful is returned, a data/wrangled folder will be downloaded and
appear in your directory.
}
\description{
Download the specified case study wrangled data to use as you follow along the case study.
}
\details{
This function downloads the Open Case Study wrangled data
from GitHub and saves it in a new 'data/wrangled/' folder in
the specified directory. This makes it so all the wrangled data
are easily available in a local folder to be analyzed.
}
\examples{
load_wrangled_data('ocs-bp-co2-emissions')

}
